Fix race condition in check_new_key
Fix a race condition where one thread could replace the inner ReentrantMutex with another one.

This can happen if two threads call check_new_key at the same time, both check for the key and don't find it and then the first one to get the write lock will insert a new ReentrantMutex under that key, but once the second thread gets the write lock, it will replace that ReentrantMutex with a different one.